chapter3/4_gui_cloud.py
import feedparser, os ,asyncio
import streamlit as st
from strands import Agent, tool

os.environ['AWS_ACCESS_KEY_ID'] = st.secrets['AWS_ACCESS_KEY_ID']
os.environ['AWS_SECRET_ACCESS_KEY'] = st.secrets['AWS_SECRET_ACCESS_KEY']
os.environ['AWS_DEFAULT_REGION'] = st.secrets['AWS_DEFAULT_REGION']

@tool
def get_aws_update(service_name: str) -> list:
    feed = feedparser.parse("https://aws.amazon.com/about-aws/whats-new/recent/feed/")
    result = []
    for entry in feed.entries:
        if service_name.lower() in entry.title.lower():
            result.append({
                "published" : entry.get("publoshed","N/A"),
                "summary" : entry.get("summart", "")
            })

            if len(result) >= 3:
                break
    
    return result

agent = Agent(
    model = "us.anthropic.claude-3-7-sonnet-20250219-v1:0",
    tools = [get_aws_update]
)

st.title("AWSアップデート確認くん")
service_name = st.text_input("アップデートを知りたいAWSサービス名を入力してください：")

async def process_stream(service_name, container):
    text_holder = container.empty()
    response = ""
    prompt = f"AWSの{service_name.strip()}の最新アップデートを、日付つきで要約して。"
    
    # エージェントからのストリーミングレスポンスを処理    
    async for chunk in agent.stream_async(prompt):
        if isinstance(chunk, dict):
            event = chunk.get("event", {})

            # ツール実行を検出して表示
            if "contentBlockStart" in event:
                tool_use = event["contentBlockStart"].get("start", {}).get("toolUse", {})
                tool_name = tool_use.get("name")
                
                # バッファをクリア
                if response:
                    text_holder.markdown(response)
                    response = ""

                # ツール実行のメッセージを表示
                container.info(f"🔧 {tool_name} ツールを実行中…")
                text_holder = container.empty()
            
            # テキストを抽出してリアルタイム表示
            if text := chunk.get("data"):
                response += text
                text_holder.markdown(response)

if st.button("確認"):
    if service_name:
        with st.spinner("アップデートを確認中..."):
            container = st.container()
            asyncio.run(process_stream(service_name,container))

    
